reworking semaphore synchronization again. Giving up on getting a decent sem_timedwait approach working.
FossilOrigin-Name: 150a6949fb6af28fa9f777694573a564f04e05bee8bf59eda119eac36713f50c

+++ arcan_frameserver_backend.h
@@ -59,7 +59,7 @@ typedef struct {
 	uint16_t vskipthresh;
 
 	/* audio */
-	uint16_t samplerate;
+	unsigned samplerate;
 	uint8_t channels;
 	uint8_t format;
 	uint16_t vfthresh;
